# openapi.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Star Wars Films API",
      "version": "1.0.0",
      "description": "API for managing Star Wars films"
    },
    "servers": [
      {
        "url": "http://localhost:3000/"
      }
    ],
...
    "paths": {
      "/films": {
        "get": {
          "tags": [
            "films"
          ],
          "summary": "Get all films",
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Film"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "films"
          ],
          "summary": "Create a new film",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilmInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Film"
                  }
                }
              }
            }
          }
        }
      },
      "/films/{id}": {
        "get": {
          "tags": [
            "films"
          ],
          "summary": "Get a film by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Film"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "films"
          ],
          "summary": "Update a film by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilmInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Film"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "films"
          ],
          "summary": "Delete a film by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Successful response"
            }
          }
        }
      },
      "/people": {
        "get": {
          "tags": [
            "peoples"
          ],
          "summary": "Get all people",
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/People"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "peoples"
          ],
          "summary": "Create a new person",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeopleInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/People"
                  }
                }
              }
            }
          }
        }
      },
      "/people/{id}": {
        "get": {
          "tags": [
            "peoples"
          ],
          "summary": "Get a person by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/People"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "peoples"
          ],
          "summary": "Update a person by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeopleInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/People"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "peoples"
          ],
          "summary": "Delete a person by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Successful response"
            }
          }
        }
      },
      "/planets": {
        "get": {
          "tags": [
            "planets"
          ],
          "summary": "Get all planets",
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Planet"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "planets"
          ],
          "summary": "Create a new planet",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Planet"
                  }
                }
              }
            }
          }
        }
      },
      "/planets/{id}": {
        "get": {
          "tags": [
            "planets"
          ],
          "summary": "Get a planet by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Planet"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "planets"
          ],
          "summary": "Update a planet by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Planet"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "planets"
          ],
          "summary": "Delete a planet by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Successful response"
            }
          }
        }
      },
      "/species": {
        "get": {
          "tags": [
            "species"
          ],
          "summary": "Get all species",
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Species"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "species"
          ],
          "summary": "Create a new species",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeciesInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Species"
                  }
                }
              }
            }
          }
        }
      },
      "/species/{id}": {
        "get": {
          "tags": [
            "species"
          ],
          "summary": "Get a species by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Species"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "species"
          ],
          "summary": "Update a species by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeciesInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Species"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "species"
          ],
          "summary": "Delete a species by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Successful response"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "Film": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string"
            },
            "field": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                },
                "director": {
                  "type": "string"
                },
                "producer": {
                  "type": "string"
                },
                "opening_crawl": {
                  "type": "string"
                },
                "release_date": {
                  "type": "string",
                  "format": "date"
                },
                "episode_id": {
                  "type": "integer"
                },
                "starships": {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                "vehicles": {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                "planets": {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                "characters": {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                "species": {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                "created": {
                  "type": "string",
                  "format": "date-time"
                },
                "edited": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "model": {
              "type": "string"
            },
            "pk": {
              "type": "integer"
            }
          }
        },
        "FilmInput": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string"
            },
            "director": {
              "type": "string"
            },
            "producer": {
              "type": "string"
            },
            "opening_crawl": {
              "type": "string"
            },
            "release_date": {
              "type": "string",
              "format": "date"
            },
            "episode_id": {
              "type": "integer"
            },
            "starships": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            },
            "vehicles": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            },
            "planets": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            },
            "characters": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            },
            "species": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            },
            "created": {
              "type": "string",
              "format": "date-time"
            },
            "edited": {
              "type": "string",
              "format": "date-time"
            }
          },
          "required": [
            "title",
            "director",
            "producer",
            "opening_crawl",
            "release_date",
            "episode_id",
            "starships",
            "vehicles",
            "planets",
            "characters",
            "species",
            "created",
            "edited"
          ]
        },
        "People": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string"
            },
            "field": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "height": {
                  "type": "integer"
                },
                "mass": {
                  "type": "integer"
                },
                "hair_color": {
                  "type": "string"
                },
                "skin_color": {
                  "type": "string"
                },
                "eye_color": {
                  "type": "string"
                },
                "birth_year": {
                  "type": "string"
                },
                "gender": {
                  "type": "string"
                },
                "homeworld": {
                  "type": "integer"
                },
                "created": {
                  "type": "string",
                  "format": "date-time"
                },
                "edited": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "model": {
              "type": "string"
            },
            "pk": {
              "type": "integer"
            }
          }
        },
        "PeopleInput": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "height": {
              "type": "integer"
            },
            "mass": {
              "type": "integer"
            },
            "hair_color": {
              "type": "string"
            },
            "skin_color": {
              "type": "string"
            },
            "eye_color": {
              "type": "string"
            },
            "birth_year": {
              "type": "string"
            },
            "gender": {
              "type": "string"
            },
            "homeworld": {
              "type": "integer"
            },
            "created": {
              "type": "string",
              "format": "date-time"
            },
            "edited": {
              "type": "string",
              "format": "date-time"
            }
          },
          "required": [
            "name",
            "height",
            "mass",
            "hair_color",
            "skin_color",
            "eye_color",
            "birth_year",
            "gender",
            "homeworld",
            "created",
            "edited"
          ]
        },
        "Planet": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string"
            },
            "field": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "rotation_period": {
                  "type": "string"
                },
                "orbital_period": {
                  "type": "string"
                },
                "diameter": {
                  "type": "string"
                },
                "climate": {
                  "type": "string"
                },
                "gravity": {
                  "type": "string"
                },
                "terrain": {
                  "type": "string"
                },
                "surface_water": {
                  "type": "string"
                },
                "population": {
                  "type": "string"
                },
                "created": {
                  "type": "string",
                  "format": "date-time"
                },
                "edited": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "model": {
              "type": "string"
            },
            "pk": {
              "type": "integer"
            }
          }
        },
        "PlanetInput": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "rotation_period": {
              "type": "string"
            },
            "orbital_period": {
              "type": "string"
            },
            "diameter": {
              "type": "string"
            },
            "climate": {
              "type": "string"
            },
            "gravity": {
              "type": "string"
            },
            "terrain": {
              "type": "string"
            },
            "surface_water": {
              "type": "string"
            },
            "population": {
              "type": "string"
            },
            "created": {
              "type": "string",
              "format": "date-time"
            },
            "edited": {
              "type": "string",
              "format": "date-time"
            }
          },
          "required": [
            "name",
            "rotation_period",
            "orbital_period",
            "diameter",
            "climate",
            "gravity",
            "terrain",
            "surface_water",
            "population",
            "created",
            "edited"
          ]
        },
        "Species": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string"
            },
            "field": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "classification": {
                  "type": "string"
                },
                "eye_colors": {
                  "type": "string"
                },
                "skin_colors": {
                  "type": "string"
                },
                "language": {
                  "type": "string"
                },
                "hair_colors": {
                  "type": "string"
                },
                "average_lifespan": {
                  "type": "string"
                },
                "average_height": {
                  "type": "string"
                },
                "homeworld": {
                  "type": "integer"
                },
                "people": {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                "created": {
                  "type": "string",
                  "format": "date-time"
                },
                "edited": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "model": {
              "type": "string"
            },
            "pk": {
              "type": "integer"
            }
          }
        },
...
      }
    }
  }